refactor(suite-common): extract address package from wallet-utils
