+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/AddTokenModal.tsx
@@ -3,8 +3,9 @@ import { type ChangeEvent, useCallback, useEffect, useState } from 'react';
 import { selectSelectedAccount } from '@suite/account';
 import { events, selectDesktopAnalyticsDep } from '@suite/analytics';
 import { Translation, useTranslation } from '@suite/intl';
+import { isAddressValid } from '@suite-common/address';
 import { useServices } from '@suite-common/dependency-injection';
-import { isAddressValid, tryGetAccountIdentity } from '@suite-common/wallet-utils';
+import { tryGetAccountIdentity } from '@suite-common/wallet-utils';
 import { Input, Modal } from '@trezor/components';
 import TrezorConnect, { type TokenInfo } from '@trezor/connect';
