+++ packages/suite/src/hooks/wallet/sign-verify/useSignVerifyForm.ts
@@ -3,9 +3,9 @@ import { useController, useForm } from 'react-hook-form';
 
 import { yupResolver } from '@hookform/resolvers/yup';
 
+import { isAddressValid } from '@suite-common/address';
 import { yup } from '@suite-common/validators';
 import { type NetworkSymbol } from '@suite-common/wallet-config';
-import { isAddressValid } from '@suite-common/wallet-utils';
 
 import type { Account } from 'src/types/wallet';
