+++ packages/suite/src/views/wallet/send/Outputs/Address.tsx
@@ -4,12 +4,19 @@ import { checkAddressCheckSum, toChecksumAddress } from 'web3-utils';
 
 import { events, selectDesktopAnalyticsDep } from '@suite/analytics';
 import { useDevice } from '@suite/device';
-import { Translation, useTranslation } from '@suite/intl';
+import { Translation, type TranslationKey, useTranslation } from '@suite/intl';
 import { Labeling } from '@suite/labeling';
 import { selectIsMetadataEnabled } from '@suite/metadata';
 import { openDeferredModal } from '@suite/modal';
 import { selectIsDebugModeActive } from '@suite/settings';
 import { selectDesktopSuiteSyncInteraction } from '@suite/suite-sync';
+import {
+    type AddressCorrection,
+    autocorrectAddress,
+    isAddressDeprecated,
+    isAddressValid,
+    isTaprootAddress,
+} from '@suite-common/address';
 import { useServices } from '@suite-common/dependency-injection';
 import { getNetworkSymbolForProtocol } from '@suite-common/suite-utils';
 import { notificationsActions } from '@suite-common/toast-notifications';
@@ -18,13 +25,7 @@ import type { Output } from '@suite-common/wallet-types';
 import {
     checkIsAddressNotUsedNotChecksummed,
     convertAmountSubunitsToUnits,
-    hasBitcoinCashAddressPrefix,
-    isAddressDeprecated,
-    isAddressValid,
-    isBech32AddressUppercase,
-    isBitcoinCashAddressUppercase,
     isProgramDerivedAccount,
-    isTaprootAddress,
 } from '@suite-common/wallet-utils';
 import { Icon, IconButton, Input, Link, Row, Text } from '@trezor/components';
 import TrezorConnect from '@trezor/connect';
@@ -48,6 +49,11 @@ import { getProtocolInfo } from 'src/utils/suite/protocol';
 import { captureSentryMessage } from 'src/utils/suite/sentry';
 
 import { DevSelfAddress } from './DevSelfAddress';
+
+const autocorrectTranslationKeys: Record<NonNullable<AddressCorrection>['type'], TranslationKey> = {
+    lowercase: 'TR_CONVERTED_TO_LOWERCASE',
+    bchPrefix: 'TR_ADDED_BITCOINCASH_PREFIX',
+};
 
 type AddressProps = {
     outputId: number;
@@ -325,16 +331,13 @@ export const Address = ({ output, outputId, outputsCount }: AddressProps) => {
                     return translationString('TR_UNSUPPORTED_ADDRESS_FORMAT');
                 }
             },
-            // bech32 and CashAddr addresses are valid as uppercase but are not accepted by Trezor
-            uppercase: (value: string) => {
-                if (
-                    (networkType === 'bitcoin' && isBech32AddressUppercase(value)) ||
-                    (symbol === 'bch' && isBitcoinCashAddressUppercase(value))
-                ) {
-                    setValue(inputName, value.toLowerCase(), { shouldValidate: true });
+            addressCorrection: (value: string) => {
+                const correction = autocorrectAddress(value, symbol);
+                if (correction) {
+                    setValue(inputName, correction.corrected, { shouldValidate: true });
                     composeTransaction();
                     setAutocorrectMessageWithTimeout(
-                        translationString('TR_CONVERTED_TO_LOWERCASE'),
+                        translationString(autocorrectTranslationKeys[correction.type]),
                     );
 
                     return true;
@@ -353,19 +356,6 @@ export const Address = ({ output, outputId, outputsCount }: AddressProps) => {
                     device?.unavailableCapabilities?.taproot
                 ) {
                     return translationString('RECIPIENT_REQUIRES_UPDATE');
-                }
-            },
-            bchMissingPrefix: (value: string) => {
-                if (symbol === 'bch' && !hasBitcoinCashAddressPrefix(value)) {
-                    setValue(inputName, 'bitcoincash:' + value, {
-                        shouldValidate: true,
-                    });
-                    setAutocorrectMessageWithTimeout(
-                        translationString('TR_ADDED_BITCOINCASH_PREFIX'),
-                    );
-                    composeTransaction();
-
-                    return true;
                 }
             },
             evmChecks: async (address: string) => {
