+++ suite-common/address/src/hasBitcoinCashAddressPrefix.ts
@@ -1,0 +1,2 @@
+export const hasBitcoinCashAddressPrefix = (address: string) =>
+    /^bitcoincash:/.test(address.toLowerCase());
