+++ suite-common/address/src/index.ts
@@ -1,1 +1,9 @@
+export { autocorrectAddress, type AddressCorrection } from './autocorrectAddress';
+export { hasBitcoinCashAddressPrefix } from './hasBitcoinCashAddressPrefix';
+export { isAddressDeprecated } from './isAddressDeprecated';
+export { isAddressValid } from './isAddressValid';
+export { isBech32AddressUppercase } from './isBech32AddressUppercase';
+export { isBitcoinCashAddressUppercase } from './isBitcoinCashAddressUppercase';
+export { isTaprootAddress } from './isTaprootAddress';
+
 export * from './getFirstFreshAddress';
