+++ suite-common/address/src/isBech32AddressUppercase.ts
@@ -1,0 +1,2 @@
+export const isBech32AddressUppercase = (address: string) =>
+    /^(bc1|tb1|ltc1|vtc1)/.test(address.toLowerCase()) && /[A-Z]/.test(address);
