+++ suite-common/address/src/isBitcoinCashAddressUppercase.ts
@@ -1,0 +1,4 @@
+import { hasBitcoinCashAddressPrefix } from './hasBitcoinCashAddressPrefix';
+
+export const isBitcoinCashAddressUppercase = (address: string) =>
+    hasBitcoinCashAddressPrefix(address) && /[A-Z]/.test(address);
