+++ suite-common/address/tsconfig.json
@@ -1,5 +1,11 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [{ "path": "../wallet-types" }]
+    "references": [
+        { "path": "../wallet-config" },
+        { "path": "../wallet-types" },
+        {
+            "path": "../../packages/address-validator"
+        }
+    ]
 }
