+++ suite-common/validators/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        { "path": "../address" },
         { "path": "../wallet-config" },
         { "path": "../wallet-utils" },
         { "path": "../../packages/utils" }
