+++ suite-common/wallet-utils/package.json
@@ -14,6 +14,7 @@
     },
     "dependencies": {
         "@scure/base": "^2.0.0",
+        "@suite-common/address": "workspace:*",
         "@suite-common/calldata": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
@@ -24,7 +25,6 @@
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-constants": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
-        "@trezor/address-validator": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/blockchain-link-utils": "workspace:*",
         "@trezor/coins-solana": "workspace:*",
