+++ suite-common/wallet-utils/src/__tests__/validation.test.ts
@@ -1,99 +1,6 @@
-import {
-    hasBitcoinCashAddressPrefix,
-    isAddressDeprecated,
-    isAddressValid,
-    isBech32AddressUppercase,
-    isBitcoinCashAddressUppercase,
-    isDecimalsValid,
-    isHexValid,
-    isInteger,
-    isTaprootAddress,
-} from '../validationUtils';
+import { isDecimalsValid, isHexValid, isInteger } from '../validationUtils';
 
 describe('validation', () => {
-    // fixtures from https://github.com/trezor/trezor-address-validator/blob/master/test/wallet_address_validator.js
-    it('isAddressValid', () => {
-        // BTC valid
-        expect(isAddressValid('12QeMLzSrB8XH8FvEzPMVoRxVAzTr5XM2y', 'btc')).toEqual(true);
-        expect(isAddressValid('3FyVFsEyyBPzHjD3qUEgX7Jsn4tcHNZFkn', 'btc')).toEqual(true);
-        expect(isAddressValid('BC1QW508D6QEJXTDG4Y5R3ZARVARY0C5XW7KV8F3T4', 'btc')).toEqual(true);
-        expect(isAddressValid('bc1qafk4yhqvj4wep57m62dgrmutldusqde8adh20d', 'btc')).toEqual(true); // p2pkh
-        expect(
-            isAddressValid('bc1q6rgl33d3s9dugudw7n68yrryajkr3ha9q8q24j20zs62se4q9tsqdy0t2q', 'btc'),
-        ).toEqual(true); // p2wsh
-        expect(
-            isAddressValid('bc1p5cyxnuxmeuwuvkwfem96lqzszd02n6xdcjrs20cac6yqjjwudpxqkedrcr', 'btc'),
-        ).toEqual(true); // p2tr
-
-        // BTC invalid
-        expect(isAddressValid('mzBc4XEFSdzCDcTxAgf6EZXgsZWpztRhef', 'btc')).toEqual(false);
-        expect(isAddressValid('2MxKEf2su6FGAUfCEAHreGFQvEYrfYNHvL7', 'btc')).toEqual(false);
-        expect(isAddressValid('bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t5', 'btc')).toEqual(false);
-        expect(isAddressValid('bc1zw508d6qejxtdg4y5r3zarvaryvg6kdaj', 'btc')).toEqual(false); // p2w-unknown
-        expect(
-            isAddressValid('tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sl5k7', 'btc'),
-        ).toEqual(false); // testnet address
-        expect(isAddressValid('BC1SW50QA3JX3S', 'btc')).toEqual(false); // p2w-unknown
-        expect(isAddressValid('bc1zw508d6qejxtdg4y5r3zarvaryvaxxpcs', 'btc')).toEqual(false); // version 2
-        expect(isAddressValid('BC1SW50QGDZ25J', 'btc')).toEqual(false); // version 16
-        expect(
-            isAddressValid(
-                'bc1pw508d6qejxtdg4y5r3zarvary0c5xw7kw508d6qejxtdg4y5r3zarvary0c5xw7k7grplx',
-                'btc',
-            ),
-        ).toEqual(false); // p2w-unknown
-
-        // TEST valid
-        expect(isAddressValid('mzBc4XEFSdzCDcTxAgf6EZXgsZWpztRhef', 'test')).toEqual(true);
-        expect(isAddressValid('2MxKEf2su6FGAUfCEAHreGFQvEYrfYNHvL7', 'test')).toEqual(true);
-        expect(
-            isAddressValid(
-                'tb1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3q0sl5k7',
-                'test',
-            ),
-        ).toEqual(true);
-        expect(
-            isAddressValid(
-                'tb1qusxlgq9quu27ucxs7a2fg8nv0pycdzvxsjk9npyupupxw3y892ssaskm8v',
-                'test',
-            ),
-        ).toEqual(true); // p2wsh
-        expect(
-            isAddressValid(
-                'tb1pn2d0yjeedavnkd8z8lhm566p0f2utm3lgvxrsdehnl94y34txmts5s7t4c',
-                'test',
-            ),
-        ).toEqual(true); // taproot
-        expect(isAddressValid('GSa5espVLNseXEfKt46zEdS6jrPkmFghBU', 'test')).toEqual(true); // regtest
-
-        // TEST invalid
-        expect(isAddressValid('12QeMLzSrB8XH8FvEzPMVoRxVAzTr5XM2y', 'test')).toEqual(false);
-        expect(isAddressValid('3FyVFsEyyBPzHjD3qUEgX7Jsn4tcHNZFkn', 'test')).toEqual(false);
-        expect(isAddressValid('bc1zw508d6qejxtdg4y5r3zarvaryvg6kdaj', 'test')).toEqual(false);
-        expect(isAddressValid('BC1SW50QA3JX3S', 'test')).toEqual(false);
-
-        // ETH, ETC, TSEP, THOD
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF', 'eth')).toEqual(true);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF0', 'eth')).toEqual(false);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF', 'etc')).toEqual(true);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF0', 'etc')).toEqual(false);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF', 'tsep')).toEqual(true);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF0', 'tsep')).toEqual(
-            false,
-        );
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF', 'thod')).toEqual(true);
-        expect(isAddressValid('0xE37c0D48d68da5c5b14E5c1a9f1CFE802776D9FF0', 'thod')).toEqual(
-            false,
-        );
-
-        // XRP, tXRP
-        expect(isAddressValid('rG1QQv2nh2gr7RCZ1P8YYcBUKCCN633jCn', 'xrp')).toEqual(true);
-        expect(isAddressValid('rG1QQv2nh2gr7RCZ1P8YYcBUKCCN633jCn', 'txrp')).toEqual(true);
-        expect(isAddressValid('rDTXLQ7ZKZVKz33zJbHjgVShjsBnqMBhmN', 'xrp')).toEqual(true);
-        expect(isAddressValid('rDTXLQ7ZKZVKz33zJbHjgVShjsBnqMBhmN', 'txrp')).toEqual(true);
-        expect(isAddressValid('r3kmLJN5D28dHuH8vZNUZpMC43pEHpaocV0', 'xrp')).toEqual(false);
-    });
-
     it('isDecimalsValid', () => {
         expect(isDecimalsValid('0', 18)).toBe(true);
         expect(isDecimalsValid('0.0', 18)).toBe(true);
@@ -141,76 +48,6 @@ describe('validation', () => {
         expect(isInteger('a01')).toBe(false);
         expect(isInteger('0a1')).toBe(false);
         expect(isInteger('01a')).toBe(false);
-    });
-
-    it('isTaprootAddress', () => {
-        expect(isTaprootAddress('', 'btc')).toBe(false);
-        expect(isTaprootAddress('bc1zw508d6qejxtdg4y5r3zarvaryvg6kdaj', 'btc')).toBe(false);
-        expect(
-            isTaprootAddress(
-                'bc1q6rgl33d3s9dugudw7n68yrryajkr3ha9q8q24j20zs62se4q9tsqdy0t2q',
-                'btc',
-            ),
-        ).toBe(false);
-        expect(
-            isTaprootAddress(
-                'bc1p5cyxnuxmeuwuvkwfem96lqzszd02n6xdcjrs20cac6yqjjwudpxqkedrcr',
-                'btc',
-            ),
-        ).toBe(true);
-        expect(
-            isTaprootAddress(
-                'tb1pn2d0yjeedavnkd8z8lhm566p0f2utm3lgvxrsdehnl94y34txmts5s7t4c',
-                'test',
-            ),
-        ).toEqual(true);
-    });
-
-    it.each([
-        ['', false],
-        ['bitcoincash', false],
-        ['bitcoincash:', true],
-        ['bitcoincash:qz8gjexl9x7gag53xl08mt7qskvjg8x2wueejjmttc', true],
-        ['BITCOINCASH:QZ8GJEXL9X7GAG53XL08MT7QSKVJG8X2WUEEJJMTTC', true],
-        ['somethingbitcoincash:something', false],
-    ])('hasBitcoinCashAddressPrefix', (address, expected) => {
-        expect(hasBitcoinCashAddressPrefix(address)).toBe(expected);
-    });
-
-    it.each([
-        ['', false],
-        ['bitcoincash:qz8gjexl9x7gag53xl08mt7qskvjg8x2wueejjmttc', false],
-        ['BITCOINCASH:QZ8GJEXL9X7GAG53XL08MT7QSKVJG8X2WUEEJJMTTC', true],
-        ['1BpEi6DfDAUFd7GtittLSdBeYJvcoaVggu', false],
-        ['1BPEI6DFDAUFD7GTITTLSDBEYJVCOAVGGU', false],
-        ['NotValidAddressContainsBITCOINCASH:BCHTEST:1', false],
-    ])('isBitcoinCashAddressUppercase', (address, expected) => {
-        expect(isBitcoinCashAddressUppercase(address)).toBe(expected);
-    });
-
-    it('isBech32AddressUppercase', () => {
-        expect(isBech32AddressUppercase('')).toBe(false);
-        expect(isBech32AddressUppercase('bc1zw508d6qejxtdg4y5r3zarvaryvg6kdaj')).toBe(false);
-        expect(isBech32AddressUppercase('BC1SW50QA3JX3S')).toBe(true);
-        expect(isBech32AddressUppercase('tb1qkvwu9g3k2pdxewfqr7syz89r3gj557l3uuf9r9')).toBe(false);
-        expect(isBech32AddressUppercase('TB1QKVWU9G3K2PDXEWFQR7SYZ89R3GJ557L3UUF9R9')).toBe(true);
-        expect(isBech32AddressUppercase('ltc1qkzyarpkhdecu5rzeuj78pwpr5sfm798afny4n6')).toBe(false);
-        expect(isBech32AddressUppercase('LTC1QKZYARPKHDECU5RZEUJ78PWPR5SFM798AFNY4N6')).toBe(true);
-        expect(isBech32AddressUppercase('37VJHKeBA9DHKmTwYE7TWYjwDzo5JTb1sz')).toBe(false); // real btc address contains tb1 string
-        expect(isBech32AddressUppercase('NotValidAddressContainsTb1bc1Ltc1Tltc1')).toBe(false);
-    });
-
-    // https://litecoin-project.github.io/p2sh-convert/
-    // https://cashaddr.bitcoincash.org/
-    it('isAddressDeprecated', () => {
-        expect(isAddressDeprecated('3notValid', 'ltc')).toBe(undefined);
-        expect(isAddressDeprecated('3NP9U8dbNzBcwhChpX8nk4F3Bf2oSucXj1', 'ltc')).toBe(
-            'LTC_ADDRESS_INFO_URL',
-        );
-        expect(isAddressDeprecated('1notValid', 'bch')).toBe(undefined);
-        expect(isAddressDeprecated('12QeMLzSrB8XH8FvEzPMVoRxVAzTr5XM2y', 'bch')).toBe(
-            'HELP_CENTER_CASHADDR_URL',
-        );
     });
 
     it('isHexValid', () => {
