+++ suite-common/wallet-utils/src/validationUtils.ts
@@ -2,73 +2,17 @@ import { type UseFormSetValue } from 'react-hook-form';
 
 import { toChecksumAddress } from 'web3-utils';
 
-import { getTestnetSymbols } from '@suite-common/wallet-config';
-import { type Account } from '@suite-common/wallet-types';
-import { type NetworkEnvironment, getAddressType, validate } from '@trezor/address-validator';
 import { type AccountInfo } from '@trezor/blockchain-link-types';
 
-const getNetworkEnvironment = (symbol: Account['symbol'], address: string): NetworkEnvironment => {
-    if (symbol === 'regtest') return symbol;
-    const testnets = getTestnetSymbols();
-
-    if (symbol === 'ada' && address.startsWith('stake')) return 'stake';
-
-    return testnets.includes(symbol) ? 'testnet' : 'prod';
-};
-
-const getCoinFromTestnet = (symbol: Account['symbol']) => {
-    switch (symbol) {
-        case 'test':
-        case 'regtest':
-            return 'btc';
-        case 'txrp':
-            return 'xrp';
-        case 'txlm':
-            return 'xlm';
-        case 'dsol':
-            return 'sol';
-        case 'tsep':
-        case 'thod':
-            return 'eth';
-        default:
-            return symbol;
-    }
-};
-
-export const isAddressValid = (address: string, symbol: Account['symbol']) => {
-    const network = getNetworkEnvironment(symbol, address);
-    const updatedSymbol = getCoinFromTestnet(symbol);
-
-    return validate(address, updatedSymbol.toUpperCase(), network);
-};
-
-export const isAddressDeprecated = (address: string, symbol: Account['symbol']) => {
-    // catch deprecated address formats
-    // LTC starting with "3" and valid with a BTC format
-    if (symbol === 'ltc' && address.startsWith('3') && isAddressValid(address, 'btc')) {
-        return 'LTC_ADDRESS_INFO_URL';
-    }
-    // BCH starting with "1" and valid with a BTC format
-    if (symbol === 'bch' && address.startsWith('1') && isAddressValid(address, 'btc')) {
-        return 'HELP_CENTER_CASHADDR_URL';
-    }
-};
-
-export const isTaprootAddress = (address: string, symbol: Account['symbol']) => {
-    const network = getNetworkEnvironment(symbol, address);
-    const updatedSymbol = getCoinFromTestnet(symbol);
-
-    return getAddressType(address, updatedSymbol.toUpperCase(), network) === 'p2tr';
-};
-
-export const hasBitcoinCashAddressPrefix = (address: string) =>
-    /^bitcoincash:/.test(address.toLowerCase());
-
-export const isBitcoinCashAddressUppercase = (address: string) =>
-    hasBitcoinCashAddressPrefix(address) && /[A-Z]/.test(address);
-
-export const isBech32AddressUppercase = (address: string) =>
-    /^(bc1|tb1|ltc1|vtc1)/.test(address.toLowerCase()) && /[A-Z]/.test(address);
+// Re-export address functions from their new home for backwards compatibility.
+export {
+    isAddressValid,
+    isAddressDeprecated,
+    isTaprootAddress,
+    hasBitcoinCashAddressPrefix,
+    isBitcoinCashAddressUppercase,
+    isBech32AddressUppercase,
+} from '@suite-common/address';
 
 export const isDecimalsValid = (value: string, decimals: number) => {
     const DECIMALS_RE = new RegExp(
