+++ suite-common/wallet-utils/tsconfig.json
@@ -3,6 +3,7 @@
     "compilerOptions": { "outDir": "libDev" },
     "include": [".", "**/*.json"],
     "references": [
+        { "path": "../address" },
         { "path": "../calldata" },
         { "path": "../earn-staking-api" },
         { "path": "../fiat-services" },
@@ -13,9 +14,6 @@
         { "path": "../wallet-config" },
         { "path": "../wallet-constants" },
         { "path": "../wallet-types" },
-        {
-            "path": "../../packages/address-validator"
-        },
         {
             "path": "../../packages/blockchain-link-types"
         },
