+++ suite-native/module-accounts-import/package.json
@@ -18,6 +18,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@shopify/flash-list": "2.3.0",
         "@shopify/react-native-skia": "2.5.1",
+        "@suite-common/address": "workspace:*",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
