+++ suite-native/module-accounts-import/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        { "path": "../../suite-common/address" },
         {
             "path": "../../suite-common/dependency-injection"
         },
