+++ suite-native/module-send/package.json
@@ -17,6 +17,7 @@
         "@react-navigation/native-stack": "7.12.0",
         "@reduxjs/toolkit": "2.11.2",
         "@shopify/flash-list": "2.3.0",
+        "@suite-common/address": "workspace:*",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
         "@suite-common/formatters": "workspace:*",
