+++ suite-native/module-send/src/sendOutputsFormSchema.ts
@@ -1,15 +1,17 @@
+import {
+    isAddressDeprecated,
+    isAddressValid,
+    isBech32AddressUppercase,
+    isTaprootAddress,
+} from '@suite-common/address';
 import { formInputsMaxLength, yup } from '@suite-common/validators';
 import { type NetworkSymbol, getDisplaySymbol, getNetworkType } from '@suite-common/wallet-config';
 import { U_INT_32 } from '@suite-common/wallet-constants';
 import { type FeeInfo, type Output } from '@suite-common/wallet-types';
 import {
     formatNetworkAmount,
-    isAddressDeprecated,
-    isAddressValid,
     isAmountWithinNetworkReserve,
-    isBech32AddressUppercase,
     isDecimalsValid,
-    isTaprootAddress,
 } from '@suite-common/wallet-utils';
 import { type FeeLevelsMaxAmount } from '@suite-native/transaction-management';
 import { BigNumber, isNotNullOrUndefined } from '@trezor/utils';
