+++ suite/e2e/package.json
@@ -26,6 +26,7 @@
         "@playwright/browser-firefox": "1.58.2",
         "@playwright/browser-webkit": "1.58.2",
         "@playwright/test": "1.58.2",
+        "@suite-common/address": "workspace:*",
         "@suite-common/analytics": "workspace:*",
         "@suite-common/device": "workspace:*",
         "@suite-common/e2e-evolu-client": "workspace:*",
